+++ examples/src/bin/echo_client.rs
@@ -34,6 +34,10 @@ use {
 // Let's set up the app.
 
 fn main() -> AppExit {
+    #[cfg(not(target_family = "wasm"))]
+    rustls::crypto::ring::default_provider()
+        .install_default()
+        .expect("ring should be the first rustls crypto provider installed");
     App::new()
         .add_plugins((
             DefaultPlugins,