+++ examples/src/bin/echo_server.rs
@@ -38,6 +38,9 @@ use {
 // Let's set up the app.
 
 fn main() -> AppExit {
+    rustls::crypto::ring::default_provider()
+        .install_default()
+        .expect("ring should be the first rustls crypto provider installed");
     App::new()
         .add_plugins((
             // Core Bevy plugins.